made monolith into components. Tested it in thunderclient. Able to sign up and sign in
